# api/docs/swagger.json
{
    "schemes": [
        "http"
    ], 
    "swagger": "2.0",
    "info": {
        "description": "The public Nebula Labs API for access to pertinent UT Dallas data",
        "title": "nebula-api",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "nebula-api-2lntm5dxoflqn.apigateway.nebula-api-368223.cloud.goog",
    "paths": {
...
        "/grades/overall": {
            "get": {
                "description": "\"Returns the overall grade distribution\"",
                "produces": [
                    "application/json"
                ],
                "operationId": "gradeAggregationOverall",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The course's subject prefix",
                        "name": "prefix",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The course's official number",
                        "name": "number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The professor's first name",
                        "name": "first_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The professors's last name",
                        "name": "last_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The number of the section",
                        "name": "section_number",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A grade distribution array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
        "/grades/semester": {
            "get": {
                "description": "\"Returns grade distributions aggregated by semester\"",
                "produces": [
                    "application/json"
                ],
                "operationId": "gradeAggregationBySemester",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The course's subject prefix",
                        "name": "prefix",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The course's official number",
                        "name": "number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The professor's first name",
                        "name": "first_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The professors's last name",
                        "name": "last_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The number of the section",
                        "name": "section_number",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of grade distributions for each semester included",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.GradeResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "responses.GradeResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    },
...
}